Fixed issue where classes would fill  up notification when they have multiple days. 'already enrolled' validation happens in cart now. little bit of refactoring on checkoutservlet
