+++ MobileEnrollment/src/java/servlets/CheckoutServlet.java
@@ -5,6 +5,7 @@
  */
 package servlets;
 
+import business.CartUtil;
 import business.Enroll;
 import business.EnrollDB;
 import business.HibernateUtil;
@@ -19,9 +20,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.hibernate.Query;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
 
 /**
  *
@@ -32,108 +30,50 @@ public class CheckoutServlet extends HttpServlet {
             throws ServletException, IOException {
         String URL = "/Cart.jsp";
         String msg = "";
-        Students s = new Students();
-        Enroll enr = new Enroll();
+        Students s = null;
+        Enroll enr = null;
         List<Section> secs = null;
         List<Enroll> enrollList = new ArrayList();
         
-        
-        
         try {
-            secs = (List<Section>)request.getSession().getAttribute("cartSections");
-            s = (Students)request.getSession().getAttribute("s");
-            String stu_id = s.getStuId();
+            secs = (List<Section>) request.getSession().getAttribute("cartSections");
+            s = (Students) request.getSession().getAttribute("s");
+            String stuId = s.getStuId();
 
             
-            //Jon's cart date/time validation
-            //this should check if any classes in the cart happen at the same time
-            for (int j = 0; j < secs.size() - 1; j++) {
-                if (secs.get(j).getDays() != null) {
-                    for (int k = j + 1; k < secs.size(); k++) {
-                        if (secs.get(k).getDays() != null) {
-                            char[] days1 = secs.get(j).getDays().toCharArray();
-                            char[] days2 = secs.get(k).getDays().toCharArray();
+            for (Section i : secs) {
+                enr = new Enroll();
+                enr.setCrn(i.getCrn());
+                enr.setStuId(s.getStuId());
+                enr.setEnrollDate(new Date());
+                enr.setSection(i);
+                enrollList.add(enr);
+            }
 
-                            for (char c:days1) {
-                                for (char c2:days2) {
-                                    if (c == c2) {
-                                        if (isOverlapping(secs.get(j).getStartTime(), secs.get(j).getEndTime(), secs.get(k).getStartTime(), secs.get(k).getEndTime())) {
-                                            msg += secs.get(j).getCrn() + " " + secs.get(j).getCourse().getCourseName() + " occurs at the same time as " + secs.get(k).getCrn() + " " + secs.get(k).getCourse().getCourseName() + "<br>";
-                                            URL = "/Cart.jsp";
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
+            for (Enroll e : enrollList) {
+                if (EnrollDB.isRegistered(e)) {
+                    msg += "You are already enrolled for " + e.getSection().getCourse().getCourseName() + "! <br><br>";
+                }
+                
+                if (!msg.isEmpty()) {
+                    msg += "Please fix any cart conflicts and retry.";
+                }
+            }
+            
+            if (msg.isEmpty()) {
+                if (!CartUtil.checkCartOverlap(secs).equals("")) {
+                    msg += CartUtil.checkCartOverlap(secs) + "<br>";
+                }
+                if (!EnrollDB.isOverlappingDB(secs, s.getStuId()).equals("")) {
+                    msg += EnrollDB.isOverlappingDB(secs, stuId);
+                }
+                
+                if (!msg.isEmpty()) {
+                    msg += "Please fix any cart conflicts and retry.";
                 }
             }
 
-                //Jon's date/time validation with the database
-                //this compares classes in the cart to already-registered classes
-                //grabbing the sessionfactory object from our hibernate utility
-                SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
-                Session session = null;
-                List<Section> sections;
-                List<Enroll> enrolled;
-
-                try {
-                    session = sessionFactory.openSession();
-                    String qs = "from Enroll where stuId = :stuId";
-                    Query q = session.createQuery(qs);
-                    q.setString("stuId", stu_id);
-
-                    enrolled = q.list();
-
-                    if (enrolled != null) {
-
-
-                        for (int j = 0; j < secs.size(); j++) {
-                            if (secs.get(j).getDays()!= null) {
-                                    for (int k = 0; k < enrolled.size(); k++) {
-                                        if (enrolled.get(k).getSection().getDays() != null) {
-                                        char[] days1 = secs.get(j).getDays().toCharArray();
-                                        char[] days2 = enrolled.get(k).getSection().getDays().toCharArray();
-
-                                        for (char c:days1) {
-                                            for (char c2:days2) {
-                                                if (c == c2) {
-                                                    if (isOverlapping(secs.get(j).getStartTime(), secs.get(j).getEndTime(), enrolled.get(k).getSection().getStartTime(), enrolled.get(k).getSection().getEndTime())) {
-                                                        msg += secs.get(j).getCrn() + " " + secs.get(j).getCourse().getCourseName() + " occurs at the same time as your previously enrolled class, " + enrolled.get(k).getCrn() + " " + enrolled.get(k).getSection().getCourse().getCourseName() + "<br>";
-                                                        URL = "/Cart.jsp";
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                } catch(Exception e) {
-                    msg = "CheckoutServlet data checking error: " + e.getMessage();
-                    System.out.println("EnrollDB data checking error: " + e.getMessage());
-                } finally {
-                    session.close();
-                }
-
-                if (msg.isEmpty()) {
-                    for (Section i : secs) {
-                        enr = new Enroll();
-                        enr.setCrn(i.getCrn());
-                        enr.setStuId(s.getStuId());
-                        enr.setEnrollDate(new Date());
-                        enr.setSection(i);
-                        enrollList.add(enr);
-                    }
-                    
-                    for (Enroll e : enrollList) {
-                        if (EnrollDB.isRegistered(enr)) {
-                            msg += "You are already enrolled for " + enr.getSection().getCourse().getCourseName() + "!";
-                        }
-                    }
-                }
-
+            
         } catch (Exception e) {
             msg += "Error on Checkout Servlet: " + e.getMessage();
         }
@@ -144,13 +84,13 @@ public class CheckoutServlet extends HttpServlet {
                 for (Enroll e : enrollList) {
                     boolean enrAdded = EnrollDB.persistSection(e);
                     if (enrAdded) {
-                        msg += "CRN " + e.getCrn() + " Registered.<br>"; 
+                        msg += e.getSection().getCourse().getCourseName() + " Registered!<br><br>"; 
                     } else {
-                        msg += "Already registered for CRN " + e.getCrn() + ".<br>";
+                        msg += "Error registering for " + e.getSection().getCourse().getCourseName() + ".<br>";
                     }
                 }
-                request.getSession().setAttribute("enrollList", enrollList);
                 
+                request.getSession().setAttribute("enrollList", enrollList);
                 URL = "/StudentHub.jsp";
                 request.getSession().setAttribute("cartSections", new ArrayList());
             } 
@@ -163,12 +103,6 @@ public class CheckoutServlet extends HttpServlet {
         disp.forward(request, response);        
         
     }
-    
-    
-    public static boolean isOverlapping(Date start1, Date end1, Date start2, Date end2) {
-        return start1.before(end2) && start2.before(end1);
-    }
-
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
      * Handles the HTTP <code>GET</code> method.
